feat: qmc5883

+++ Drone/Core/Inc/SensorDriver/BMI088/bmi088.h
@@ -7,7 +7,7 @@
 #define BMI088_G			9.80665f
 #define DEG2RAD				0.017453292519943295f
 
-#define BMI088_CALIB_DEFAULT_SAMPLES    10000
+#define BMI088_CALIB_DEFAULT_SAMPLES	10000
 #define BMI088_CALIB_GYRO_VAR_LIMIT     0.005f
 
 typedef struct{
@@ -27,6 +27,6 @@ BMI088_Status_t BMI088_ReadGyro(void);
 BMI088_Status_t BMI088_GetData(BMI088_Data_t *imu);
 void BMI088_ParseData(void);
 void BMI088_Convert(void);
-BMI088_Status_t BMI088_Calibrate(uint16_t numSamples);
+BMI088_Status_t BMI088_Calibrate(uint32_t numSamples);
 
 #endif /* INC_SENSORDRIVER_BMI088_BMI088_H_ */
